+++ index.php
@@ -11,14 +11,20 @@
   <link rel="stylesheet" href="res/main.css">
   <link rel="stylesheet" href="res/jquery-autocomplete.css">
 </head>
-<body data-spy="scroll" data-target="#row-pages" style="position:relative">
-<nav class="navbar navbar-dark bg-inverse navbar-fixed-top">
-  <a href="#top" class="navbar-brand hidden-xs-down">CoreProtect Lookup Web Interface</a>
-  <ul id="row-pages" class="nav navbar-nav">
-  </ul>
-  <?=$loginRequired?'<a href="index.php?action=clear_login" class="btn btn-secondary-outline pull-xs-right">logout</a>':""?>
+<body data-spy="scroll" data-target="#row-pages">
+<nav id="top" class="navbar navbar-light bg-faded navbar-full">
+  <div class="container">
+    <h1 class="navbar-brand">CoreProtect Lookup Web Interface</h1>
+    <?=$loginRequired?'<a href="index.php?action=clear_login" class="btn btn-info-outline pull-xs-right">logout</a>':""?>
+  </div>
 </nav>
-<div id="top" style="height:4rem;"></div>
+<nav class="navbar navbar-dark bg-inverse navbar-fixed-bottom">
+  <div class="container-fluid">
+    <ul id="row-pages" class="nav navbar-nav">
+      <li class="nav-item"><a class="nav-link" href="#top">Top</a></li>
+    </ul>
+  </div>
+</nav>
 <div class="container">
 <div id="lookupForm" class="card">
 <div class="card-header"><span class="h4 card-title">Make a Lookup</span></div>
@@ -26,12 +32,12 @@
 <div class="form-group row">
   <div class="col-lg-2 form-control-label">Actions</div>
   <div class="dtButtons btn-group col-lg-10">
-    <label class="btn btn-secondary" for="abl"><input type="checkbox" id="abl" name="a[]" value="block"<?=(!$fm||in_array("blocK",$_GET["a"]))?" checked":""?>>Block</label>
+    <label class="btn btn-secondary" for="abl"><input type="checkbox" id="abl" name="a[]" value="block"<?=(!$fm||in_array("block",$_GET["a"]))?" checked":""?>>Block</label>
     <label class="btn btn-secondary" for="acl"><input type="checkbox" id="acl" name="a[]" value="click"<?=($fm&&in_array("click",$_GET["a"]))?" checked":""?>>Click</label>
     <label class="btn btn-secondary" for="acn"><input type="checkbox" id="acn" name="a[]" value="container"<?=($fm&&in_array("container",$_GET["a"]))?" checked":""?>>Container</label>
     <label class="btn btn-secondary" for="ach"><input type="checkbox" id="ach" name="a[]" value="chat"<?=($fm&&in_array("chat",$_GET["a"]))?" checked":""?>>Chat</label>
     <label class="btn btn-secondary" for="acm"><input type="checkbox" id="acm" name="a[]" value="command"<?=($fm&&in_array("command",$_GET["a"]))?" checked":""?>>Command</label>
-    <label class="btn btn-secondary" for="aki"><input type="checkbox" id="akl" name="a[]" value="kill"<?=($fm&&in_array("kill",$_GET["a"]))?" checked":""?>>Kill</label>
+    <label class="btn btn-secondary" for="akl"><input type="checkbox" id="akl" name="a[]" value="kill"<?=($fm&&in_array("kill",$_GET["a"]))?" checked":""?>>Kill</label>
     <label class="btn btn-secondary" for="ass"><input type="checkbox" id="ass" name="a[]" value="session"<?=($fm&&in_array("session",$_GET["a"]))?" checked":""?>>Session</label>
     <label class="btn btn-secondary" for="aus"><input type="checkbox" id="aus" name="a[]" value="username"<?=($fm&&in_array("username",$_GET["a"]))?" checked":""?>>Username</label>
   </div>
@@ -41,64 +47,64 @@
   <div class="col-lg-10">
     <button class="btn btn-secondary" type="button" id="rcToggle">Radius/Corners</button>
     <span class="dtButtons btn-group">
-    <label class="btn btn-success-outline" for="rbt"><input type="radio" id="rbt" name="rollback" value="1"><span class="glyphicon glyphicon-ok"></span></label>
-    <label class="btn btn-secondary active" for="rb"><input type="radio" id="rb" name="rollback" value="" checked>Rollback</label>
-    <label class="btn btn-secondary-outline" for="rbf"><input type="radio" id="rbf" name="rollback" value="0"><span class="glyphicon glyphicon-minus"></span></label>
+    <label class="btn btn-success-outline" for="rbt"><input type="radio" id="rbt" name="rollback" value="1"<?=($fm&&$_GET["rollback"]==="1")?" checked":""?>><span class="glyphicon glyphicon-ok"></span></label>
+    <label class="btn btn-secondary active" for="rb"><input type="radio" id="rb" name="rollback" value=""<?=(!$fm||$_GET["rollback"]==="")?" checked":""?>>Rollback</label>
+    <label class="btn btn-secondary-outline" for="rbf"><input type="radio" id="rbf" name="rollback" value="0"<?=($fm&&$_GET["rollback"]==="0")?" checked":""?>><span class="glyphicon glyphicon-minus"></span></label>
     </span>
   </div>
 </div>
 <div class="form-group row">
     <label class="col-sm-2 form-control-label" for="x1" id="corner1">Center / Corner 1</label>
     <div class="input-group col-lg-4 col-sm-10 groups-line" id="c1">
-      <input class="form-control" type="number" id="x1" name="xyz[]" placeholder="x">
+      <input class="form-control" type="number" id="x1" name="xyz[]" placeholder="x"<?=$fm?' value="'.$_GET["xyz"][0].'"':""?>>
         <span class="input-group-btn" style="width:0"></span>
-      <input class="form-control" type="number" id="y1" name="xyz[]" placeholder="y">
+      <input class="form-control" type="number" id="y1" name="xyz[]" placeholder="y"<?=$fm?' value="'.$_GET["xyz"][1].'"':""?>>
         <span class="input-group-btn" style="width:0"></span>
-      <input class="form-control" type="number" id="z1" name="xyz[]" placeholder="z">
+      <input class="form-control" type="number" id="z1" name="xyz[]" placeholder="z"<?=$fm?' value="'.$_GET["xyz"][2].'"':""?>>
     </div>
     <label class="col-sm-2 form-control-label" for="x2" id="corner2">Radius / Corner 2</label>
     <div class="input-group col-lg-4 col-sm-10" id="c2">
-      <input class="form-control" type="number" id="x2" name="xyz2[]" placeholder="Radius or x">
+      <input class="form-control" type="number" id="x2" name="xyz2[]" placeholder="Radius or x"<?=$fm?' value="'.$_GET["xyz2"][0].'"':""?>>
       <span class="input-group-btn c2" style="width:0"></span>
-      <input class="form-control c2" type="number" id="y2" name="xyz2[]" placeholder="y">
+      <input class="form-control c2" type="number" id="y2" name="xyz2[]" placeholder="y"<?=$fm?' value="'.$_GET["xyz2"][1].'"':""?>>
       <span class="input-group-btn c2" style="width:0"></span>
-      <input class="form-control c2" type="number" id="z2" name="xyz2[]" placeholder="z">
+      <input class="form-control c2" type="number" id="z2" name="xyz2[]" placeholder="z"<?=$fm?' value="'.$_GET["xyz2"][2].'"':""?>>
     </div>
 </div>
 <div class="form-group row">
   <label class="col-xs-2 form-control-label" for="wid">World</label>
-  <div class="col-xs-10"><input class="form-control autocomplete" data-qftr="world" type="text" id="wid" name="wid" placeholder="world"></div>
+  <div class="col-xs-10"><input class="form-control autocomplete" data-qftr="world" type="text" id="wid" name="wid" placeholder="world"<?=$fm?' value="'.$_GET["wid"].'"':""?>></div>
 </div>
 <div class="form-group row">
   <label class="col-lg-2 form-control-label" for="usr">Users</label>
   <div class="input-group col-lg-10" >
-    <span class="dtButtons input-group-btn"><label class="btn btn-secondary" for="eus"><input type="checkbox" id="eus" name="e[]" value="u">Exclude</label></span>
-    <input class="form-control autocomplete" data-qftr="user" type="text" pattern="((#[a-zA-Z_]+)|([a-zA-Z0-9_]{2,16}))(,\s?((#[a-zA-Z_]+)|([a-zA-Z0-9_]{2,16})))*" id="usr" name="u" placeholder="Separate by single comma(,)">
+    <span class="dtButtons input-group-btn"><label class="btn btn-secondary" for="eus"><input type="checkbox" id="eus" name="e[]" value="u"<?=($fm&&in_array("u",$_GET["e"]))?" checked":""?>>Exclude</label></span>
+    <input class="form-control autocomplete" data-qftr="user" type="text" pattern="((#[a-zA-Z_]+)|([a-zA-Z0-9_]{2,16}))(,\s?((#[a-zA-Z_]+)|([a-zA-Z0-9_]{2,16})))*" id="usr" name="u" placeholder="Separate by single comma(,)"<?=$fm?' value="'.$_GET["u"].'"':""?>>
   </div>
 </div>
 <div class="form-group row">
   <label class="col-lg-2 form-control-label" for="blk">Blocks</label>
   <div class="input-group col-lg-10">
-    <span class="dtButtons input-group-btn"><label class="btn btn-secondary" for="ebl"><input type="checkbox" id="ebl" name="e[]" value="b">Exclude</label></span>
-    <input class="form-control autocomplete" data-qftr="material" type="text" pattern="([^:]+:[^:,]+)+" id="blk" name="b" placeholder="minecraft:<block> - Separate by single comma(,)">
+    <span class="dtButtons input-group-btn"><label class="btn btn-secondary" for="ebl"><input type="checkbox" id="ebl" name="e[]" value="b"<?=($fm&&in_array("b",$_GET["e"]))?" checked":""?>>Exclude</label></span>
+    <input class="form-control autocomplete" data-qftr="material" type="text" pattern="([^:]+:[^:,]+)+" id="blk" name="b" placeholder="minecraft:<block> - Separate by single comma(,)"<?=$fm?' value="'.$_GET["b"].'"':""?>>
   </div>
 </div>
 <div class="form-group row">
   <label class="col-sm-2 form-control-label" for="kwd">Keyword</label>
-  <div class="col-sm-10"><input class="form-control" type="text" id="kwd" name="keyword"></div>
+  <div class="col-sm-10"><input class="form-control" type="text" id="kwd" name="keyword"<?=$fm?' value="'.$_GET["keyword"].'"':""?>></div>
 </div>
 <div class="form-group row">
   <label class="col-sm-2 form-control-label" for="date">Date/Time</label>
   <div class="input-group col-lg-4 col-sm-10 groups-line">
     <span class="dtButtons input-group-btn">
-      <label class="btn btn-secondary" for="trv"><input type="checkbox" id="trv" name="asendt">Reverse</label>
+      <label class="btn btn-secondary" for="trv"><input type="checkbox" id="trv" name="asendt"<?=($fm&&$_GET["asendt"]=="on")?" checked":""?>>Reverse</label>
     </span>
-    <input class="form-control" type="datetime-local" id="date" name="t" placeholder="0000-00-00T00:00:00">
+    <input class="form-control" type="datetime-local" id="date" name="t" placeholder="0000-00-00T00:00:00"<?=$fm?' value="'.$_GET["t"].'"':""?>>
   </div>
   <input type="hidden" name="unixtime" value="on">
   <label class="col-sm-2 form-control-label" for="lim">Limit</label>
   <div class="col-lg-4 col-sm-10">
-    <input class="form-control" type="number" id="lim" name="lim" min="1" placeholder="30">
+    <input class="form-control" type="number" id="lim" name="lim" min="1" placeholder="30"<?=$fm?' value="'.$_GET["lim"].'"':""?>>
   </div>
 </div>
 <div class="row">
@@ -113,9 +119,9 @@
 <table id="output" class="table table-sm table-striped">
   <caption id="genTime"></caption>
   <thead class="thead-inverse">
-  <tr><th>Time</th><th>User</th><th>Action</th><th>Coordinates / World</th><th>Block/Item:Data</th><th>Amount</th><th>Rollback</th></tr>
+  <tr id="row-0"><th>#</th><th>Time</th><th>User</th><th>Action</th><th>Coordinates / World</th><th>Block/Item:Data</th><th>Amount</th><th>Rollback</th></tr>
   </thead>
-  <tbody id="mainTbl"><tr><td colspan="7">Please submit a lookup.</td></tr></tbody>
+  <tbody id="mainTbl"><?=isset($mainTbl)?$mainTbl:'<tr><th scope="row">-</th><td colspan="7">Please submit a lookup.</td></tr>'?></tbody>
 </table>
 </div>
 <form class="container" id="loadMore" action="">
@@ -153,6 +159,8 @@ $dynmapURL = "<?=$dynmapURL?>";
 $dynmapZoom = "<?=$dynmapZoom?>";
 $dynmapMapName = "<?=$dynmapMapName?>";
 $pageInterval = <?=$pageInteval?>;
+$fm = <?=$fm?"true":"false"?>;
+$PHP_$t = <?=($fm&&$_GET["t"]!=="")?' value="'.$_GET["t"].'"':"false"?>;
 </script>
 <script src="//ajax.googleapis.com/ajax/libs/jquery/2.2.0/jquery.min.js"></script>
 <script src="//code.jquery.com/ui/1.11.4/jquery-ui.js"></script>
@@ -164,7 +172,7 @@ $pageInterval = <?=$pageInteval?>;
 <script src="res/form-handler.js"></script>
 <div class="container">
 <p>If you encounter any issues, please open an issue or a ticket on the <a href="https://github.com/SimonOrJ/CoreProtect-Lookup-Web-Interface">GitHub project page</a> or the <a href="http://dev.bukkit.org/bukkit-plugins/coreprotect-lwi/">Bukkit plugin project page</a>.</p>
-<p>&copy; <?=$copyright?> &mdash; CoreProtect LWI version 0.7.0-beta<br>Created by <a href="http://simonorj.com/">SimonOrJ</a>.</p>
+<p>&copy; <?=$copyright?> &mdash; CoreProtect LWI version 0.7.1-beta<br>Created by <a href="http://simonorj.com/">SimonOrJ</a>.</p>
 </div>
 </body>
 </html>